: use float for time and duration because of / operator

+++ internal/player/wav_player.go
@@ -25,10 +25,10 @@ func (p *WAVPlayer) Play(irp *ir.Program, start, end int) error {
 	// ts := ipr.TimeScale()
 	instructions := irp.Instructions()
 	duration := irp.Duration()
-	totalSamples := duration * sr
+	totalSamples := int(duration * float64(sr))
 	logs.Info("Total instructions %d", irp.Length())
 	logs.Info("Total samples %d", totalSamples)
-	logs.Info("Total duration %d", duration)
+	logs.Info("Total duration %f", duration)
 
 	logs.Debug("get samples")
 	out := p.getSamples(instructions, totalSamples, audio.WaveSaw)
@@ -60,13 +60,13 @@ func (p *WAVPlayer) Stop() error {
 func (p *WAVPlayer) getSamples2(irp *ir.Program, wave audio.WaveFunc) []int16 {
 	sr := p.wav.SampleRate()
 	timeScale := 2.0 // TODO
-	total := sr * irp.Duration()
+	total := int(float64(sr) * irp.Duration())
 	buf := make([]float64, total)
 	out := make([]int16, total)
 	clip := float64(math.MaxInt16)
 
 	for _, ins := range irp.Instructions() {
-		offset := int(float64(ins.Time)*timeScale) * sr
+		offset := int(ins.Time*timeScale) * sr
 		samples := p.wav.GenerateSamplesEnvelope(
 			ins.Midi,
 			float64(ins.Dur)*timeScale,
@@ -94,8 +94,8 @@ func (p *WAVPlayer) getSamples(instructions []*ir.Instruction, total int, wave a
 
 	forDebugTimeTempVariableAtAll := 4.0
 	for _, ins := range instructions {
-		offset := ins.Time * sr / int(forDebugTimeTempVariableAtAll)
-		dur := float64(ins.Dur) / forDebugTimeTempVariableAtAll
+		offset := int(ins.Time * float64(sr) / forDebugTimeTempVariableAtAll)
+		dur := ins.Dur / forDebugTimeTempVariableAtAll
 		// TODO
 		// samples := p.wav.GenerateSamples(ins.Freq, ins.Dur, ins.Vol, wave)
 		samples := p.wav.GenerateSamplesEnvelope(audio.MidiToFreq(int(ins.Midi)), dur, ins.Vol, 0.03, 0.06, wave)
